% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_access_utils.R
\name{extract_ma}
\alias{extract_ma}
\title{extract tables from an ms access database file (.accdb)}
\usage{
extract_ma(file)
}
\arguments{
\item{file}{(character) a file path, must exist}
}
\value{
a list of tibbles/data.frame's
}
\description{
uses mdbtools command line tools, specifically mdb-tables and mdb-export
}
\details{
mdbtools is required to run this function.
\itemize{
\item macos: brew install mdbtools
\item deb: apt-get install mdbtools
\item fedora: yum install mdbtools
\item windows: not sure, anyone?
}
}
\examples{
\dontrun{
file = "/Users/sckott/Downloads/laselva_data/spain/Sig_05.accdb"
extract_ma(file)
file = "/Users/sckott/Downloads/laselva_data/spain/Ifn3p05.accdb"
extract_ma(file)
}
}
\references{
https://github.com/brianb/mdbtools
}
\keyword{internal}
