% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_datasets.R
\name{fia_datasets}
\alias{fia_datasets}
\title{list datasets}
\usage{
fia_datasets(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a list of three data.frames (as tibble's)
}
\description{
list datasets
}
\examples{
\dontrun{
fia_datasets()
}
}

