% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_fetch.R
\name{fia_fetch}
\alias{fia_fetch}
\title{list datasets}
\usage{
fia_fetch(state, what = "tree", overwrite = FALSE, ...)
}
\arguments{
\item{state}{(character) a state, two letter code}

\item{what}{(character) what data to get, default: tree}

\item{overwrite}{(logical) Will only overwrite existing path if \code{TRUE}.
default: \code{FALSE}}

\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a tibble (a data.frame)
}
\description{
list datasets
}
\examples{
\dontrun{
fia_fetch(state = "MP")
}
}

