% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_datasets.R
\name{list_datasets_usa}
\alias{list_datasets_usa}
\title{List USA Forest Inventory and Analysis (FIA) datasets}
\usage{
list_datasets_usa(...)
}
\arguments{
\item{...}{curl options passed on to [crul::HttpClient]}
}
\value{
a list of three data.frames (as tibble's), one for
reference metadata files, datasets by state and datasets
with all states combined
}
\description{
List USA Forest Inventory and Analysis (FIA) datasets
}
\examples{
\dontrun{
x <- list_datasets_usa()
x$ref_table
x$by_state
x$all_states
}
}
\references{
<https://apps.fs.usda.gov/fia/datamart/CSV/datamart_csv.html>
}
