% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_aus.R
\name{ls_fetch_aus}
\alias{ls_fetch_aus}
\title{Queensland, Australian data}
\usage{
ls_fetch_aus(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles
\itemize{
\item AuxiliaryData.csv: Each record represents presence data for species on
each plot not recorded as a stem ≥10 cm DBH or understory individuals.
\item TreeMeasurementData.csv: Each record represents the data for one stem for
a particular year shown in the variable 'year'.
\item UnderstoryData.csv: Each record represents the data for a species in the
understory (≥50 cm in height and <10 cm DBH) in a particular plot and
subplot and the presence or absence in two size classes (≥50 cm in height
to 3 m in height, and >3 m in height to <10 cm DBH).
\item VoucherData.csv: Each record represents a voucher specimen for a species
represented either by a stem ≥10 cm DBH, an understory species, or and
auxiliary species.
}

See http://esapubs.org/archive/ecol/E095/209/metadata.php and/or
https://figshare.com/articles/Data_Paper_Data_Paper/3560154 for more
information on what all the variables mean
}
\description{
Queensland, Australian data
}
\examples{
\dontrun{
res = ls_fetch_aus()
res
}
}
\references{
https://doi.org/10.1890/14-0458R.1
http://esapubs.org/archive/ecol/E095/209/metadata.php
https://figshare.com/articles/Data_Paper_Data_Paper/3560154
}
