% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_deu.R
\name{ls_fetch_deu}
\alias{ls_fetch_deu}
\title{Germany's Second and Third National Forest Inventory data.}
\usage{
ls_fetch_deu(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles
}
\description{
Note: command line tool mdbtools required to use this function
}
\details{
a list of tibbles :
\itemize{
\item shapefile : tibble with plot informations. Loaded with {sf} package.
\item daten2002 : list of tibbles for the 2002 inventory.
\item daten2012 : list of tibbles for the 2012 inventory.
\item daten2002-2012 : list of tibbles for 2012 revisit.
}

Code first version was written by \href{https://demographicecology.com/team/lukas-heiland/}{Lukas Heiland}
Second version was written by Julien Barrere
}
\section{Command line tool required}{

mdbtools is required to run this function. installation for most systems:
\itemize{
\item macos: brew install mdbtools
\item deb: apt-get install mdbtools
\item fedora: yum install mdbtools
\item windows: not sure, anyone?
}
}

\examples{
\dontrun{
res = ls_fetch_deu()
names(res)
}

}
\references{
https://bwi.info/Download/de/BWI-Basisdaten/ACCESS2003
}
