% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_esp.R
\name{ls_fetch_esp}
\alias{ls_fetch_esp}
\title{Spain's Third Naitonal Forest Inventory data}
\usage{
ls_fetch_esp(location, ...)
}
\arguments{
\item{location}{(character) location name, must be length 1. see details.
you can pass in names with accepts or capitalized; we'll standardize
to lowercase and remove accents.}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles
}
\description{
Note: command line tool mdbtools required to use this function
}
\details{
Location name must be one of:
\itemize{
\item alava
\item albacete
\item alicante
\item almeria
\item avila
\item badajoz
\item islas_baleares
\item barcelona
\item burgos
\item caceres
\item cadiz
\item castellon
\item ciudad_real
\item cordoba
\item a_coruna
\item cuenca
\item girona
\item granada
\item guadalajara
\item guipuzcoa
\item huelva
\item huesca
\item jaen
\item leon
\item lleida
}
}
\section{Command line tool required}{

mdbtools is required to run this function. installation for most systems:
\itemize{
\item macos: brew install mdbtools
\item deb: apt-get install mdbtools
\item fedora: yum install mdbtools
\item windows: not sure, anyone?
}
}

\examples{
\dontrun{
res = ls_fetch_esp(location = "alava")
res
ls_fetch_esp(location = "Álava")
ls_fetch_esp(location = "girona")
}
}
\references{
https://www.miteco.gob.es/fr/biodiversidad/servicios/banco-datos-naturaleza/informacion-disponible/ifn3_base_datos_1_25.aspx
}
