% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_eu.R
\name{ls_fetch_eu}
\alias{ls_fetch_eu}
\title{Tree inventory data for 21 countries in Europe}
\usage{
ls_fetch_eu(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles
\itemize{
\item genus: data at the generic level
\item species: data at the species level
}
}
\description{
Tree inventory data for 21 countries in Europe
}
\section{Fields}{

\itemize{
\item \code{x}/\code{y}: coordinates in a ETRS89-LAEA reference coordinate
system, representing the centroid of the INSPIRE-compliant 1 km×1 km
European grid
\item \code{country}: the country where the forest plot was sampled
\item \code{genus_name}: the name of the genus sampled
\item \code{species_name}: the name of the species sampled
\item \code{dbh_1}/\code{dbh_2}: the class for the diameter at breast height of the
trunk, which is used to distinguish a new recruit of a given tree.
This latter measurement has a value of 1 for trunks with a diameter
of less than 120 mm, 2 for trunks with diameter greater than 120 mm,
and −9999 for trunks with unknown diameters. The value of 120 mm
represents the common minimum threshold adopted by the nfis,
although single nfis have different minimum thresholds, ranging from
0 mm in Finland to 120 mm in Cyprus and Switzerland. The choice
of threshold has important implications, since estimations of biomass
based on different thresholds can deliver very different results.
\item \code{nfi}: 1 if the data comes from the nfi, -9999 if not
\item \code{ff}: 1 if the data comes from the ff, -9999 if not
\item \code{bs}: 1 if the data comes from the bs, -9999 if not
\item \code{eeo}: whether or not the target occurrence falls within the
species’ geographical range, measured as the extent of occurrence,
1 for yes, 0 for no
}

Acronyms
\itemize{
\item bs: Biosoil
\item ff: Forest Focus monitoring database
\item nfi: National Forest Inventories dataset
}

See https://doi.org/10.1038/sdata.2016.123 for more information
}

\examples{
\dontrun{
res = ls_fetch_eu()
res
}
}
\references{
https://doi.org/10.1038/sdata.2016.123
https://doi.org/10.6084/m9.figshare.c.3288407.v1
}
