% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_fin_swe.R
\name{ls_fetch_fin_swe}
\alias{ls_fetch_fin_swe}
\title{Finland and Sweden Forest inventory}
\usage{
ls_fetch_fin_swe(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles

Trees
\itemize{
\item trees_finland_and_sweden.csv
}

Plots information
\itemize{
\item plots_finland_and_sweden.csv
}

Climate information
\itemize{
\item climate_wai_sgdd.csv
}
}
\description{
Fetch Finnish and Swedish National Forest Inventories.
}
\examples{
\dontrun{
res = ls_fetch_fin_swe()
res
}

}
\references{
https://datadryad.org/stash/dataset/doi:10.5061/dryad.wm37pvmkw
}
