% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_fr.R
\name{ls_fetch_fr}
\alias{ls_fetch_fr}
\alias{ls_fetch_fr_revisit}
\title{French IGN data}
\usage{
ls_fetch_fr(year, ...)

ls_fetch_fr_revisit(year, ...)
}
\arguments{
\item{year}{(integer/numeric) data year to get. possible years: 2005-2018}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles

Documentation
\itemize{
\item documentation_year.csv
\item documentation_flora.csv
}

Dead and windfallen trees
\itemize{
\item dead_trees_forest_year.csv
\item dead_trees_poplar_year.csv
}

Covers
\itemize{
\item cover_forest_year.csv
}

Ecology
\itemize{
\item ecology_year.csv
}

FLora
\itemize{
\item flora_year.csv
}

First visits plots
\itemize{
\item plots_forest_year.csv
\item plots_poplar_year.csv
}

Living trees
\itemize{
\item trees_forest_year.csv
\item trees_poplar_year.csv
}
}
\description{
Fetch French IGN Institut National de L'Information Geographique
et Forestiere data
}
\details{
\code{ls_fetch_fr()} gets the given years data, while
\code{ls_fetch_fr_revisit()} gets revisit data for 5 years after the supplied
year
}
\examples{
\dontrun{
res = ls_fetch_fr(year = 2017)
res
ls_fetch_fr(year = 2007)

# revisit data
ls_fetch_fr_revisit(year = 2007)
}
}
\references{
https://inventaire-forestier.ign.fr/
}
