% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_jpn.R
\name{ls_fetch_jpn}
\alias{ls_fetch_jpn}
\title{Japan's Forest Data}
\usage{
ls_fetch_jpn(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of data and metadata:
\itemize{
\item species_list: a data.frame with species information
\item site_list: a data.frame with site information
\item tree_data: a list of data.frame's named by their sites, which match the
column \code{PlotID} in the \code{site_list} data.frame (except \code{-} is swapped with
\verb{_} for easier indexing)
\item metadata: an object of class \code{emld} containing metadata for the data
}
}
\description{
Japan Ministry of the Environment, Monitoring Sites 1000 Project
}
\note{
The \code{emld} R package is required for EML metadata access. If you
don't have it installed we'll just return an empty list
}
\examples{
\dontrun{
res = ls_fetch_jpn()
res
}
}
\references{
http://db.cger.nies.go.jp/JaLTER/ER_DataPapers/archives/2011/ERDP-2011-01/metadata
}
