% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_fetch_usa.R
\name{ls_fetch_usa}
\alias{ls_fetch_usa}
\title{Fetch USA Forest Inventory and Analysis (FIA) datasets}
\usage{
ls_fetch_usa(state, what = "tree", ...)
}
\arguments{
\item{state}{(character) one or more states (2 letter code), or the special
"all", for all states}

\item{what}{(character) what data to get, one or more values, default: tree}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a tibble (a data.frame)
}
\description{
Fetch USA Forest Inventory and Analysis (FIA) datasets
}
\examples{
\dontrun{
# Northern Mariana Islands - trees
ls_fetch_usa(state = "MP")
# Guam - trees
ls_fetch_usa(state = "GU")
# Guam - seedling
ls_fetch_usa(state = "GU", what = "seedling")
# Alaska - plot
ak_plot <- ls_fetch_usa(state = "AK", what = "plot")
# Arizona - vegetiation subplot - no data, empty data.frame
ls_fetch_usa(state = "AZ", what = "veg_subplot")

# multiple states
x <- ls_fetch_usa(state = c('mp', 'gu'))
x$mp_tree
x$gu_tree

# multiple datasets
y <- ls_fetch_usa('as', what = c('seedling', 'plot'))
y$as_seedling
y$as_plot

# multiple states and multiple datasets
z <- ls_fetch_usa(state = c('mp', 'gu'), what = c('seedling', 'plot'))
z$mp_seedling
z$mp_plot
z$gu_seedling
z$gu_plot

# all states, be careful, lots of data
## ls_fetch_usa("all", "subplot_regen")
}
}
