% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_list_usa.R
\name{ls_list_usa}
\alias{ls_list_usa}
\title{List USA Forest Inventory and Analysis (FIA) datasets}
\usage{
ls_list_usa(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of three data.frames (as tibble's), one for
reference metadata files, datasets by state and datasets
with all states combined
}
\description{
List USA Forest Inventory and Analysis (FIA) datasets
}
\examples{
\dontrun{
x <- ls_list_usa()
x
x$ref_table
x$by_state
x$all_states
}
}
\references{
https://apps.fs.usda.gov/fia/datamart/CSV/datamart_csv.html
}
