% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fia_datasets.R
\name{list_datasets_usa}
\alias{list_datasets_usa}
\title{List  US FIA datasets}
\usage{
list_datasets_usa(...)
}
\arguments{
\item{...}{curl options passed on to \code{\link[httr]{GET}}}
}
\value{
a list of three data.frames (as tibble's)
}
\description{
List  US FIA datasets
}
\examples{
\dontrun{
list_datasets_usa()
}
}

