% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_fra.R
\name{ls_fetch_fra}
\alias{ls_fetch_fra}
\alias{ls_fetch_fra_revisit}
\alias{ls_fetch_fra_raw}
\title{French IGN NFI data}
\usage{
ls_fetch_fra(year, ...)

ls_fetch_fra_revisit(year, ...)

ls_fetch_fra_raw(...)
}
\arguments{
\item{year}{Deprecated argument from \code{ls_fetch_fa()} and
\code{ls_fetch_fra_revisit()}.}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles

Trees alive or dead
\itemize{
\item ARBRE.csv
}

Dead wood
\itemize{
\item BOIS_MORT.csv
}

Covers
\itemize{
\item COUVERT.csv
}

Ecology
\itemize{
\item ECOLOGIE.csv
}

Flora
\itemize{
\item FLORE.csv
}

Plots information
\itemize{
\item HABITAT.csv
\item PLACETTE.csv
}

Species key
\itemize{
\item espar-cdref13.csv
}
}
\description{
Fetch French IGN Institut National de L'Information Geographique
et Forestiere data
}
\details{
metadonnees.csv was not included in the export because of an erroneous format.
This file is available for read in the cache, find where
files are being cached with \code{laselva_cache$cache_path_get}. The path is
\code{france-ign/metadonnees.csv} from there.

Documentation is available in the same \verb{france-ign/} cache as pdf files.
This documentation is written in french.

\code{ls_fetch_fra()} and \code{ls_fetch_fra_revisit()} have been deprecated due to
an IGN download method update. These functions will trigger a
\code{ls_fetch_fra_raw()} call in place.
}
\examples{
\dontrun{
res = ls_fetch_fr_raw()
res
}
}
\references{
https://inventaire-forestier.ign.fr/
}
