% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_fetch_jpn.R
\name{ls_fetch_jpn}
\alias{ls_fetch_jpn}
\title{Japan's Forest Data}
\usage{
ls_fetch_jpn(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of tibbles
}
\description{
Japan Ministry of the Environment, Monitoring Sites 1000 Project
}
\examples{
\dontrun{
res = ls_fetch_jpn()
res
}
}
\references{
http://db.cger.nies.go.jp/JaLTER/ER_DataPapers/archives/2011/ERDP-2011-01/metadata
}
